requirements.txt
numpy
pandas
streamlit
pillow
gdown
plotly
tensorflow
#opencv-python